+++ services/ark-dashboard/ark-dashboard/components/sessions-conversations/chat-input.tsx
@@ -215,6 +215,7 @@ export function ChatInput({
               <Button
                 variant="ghost"
                 size="icon-sm"
+                aria-pressed={showToolCalls}
                 aria-label={
                   showToolCalls ? 'Hide tool calls' : 'Show tool calls'
                 }
